Clicking on the nearby FAB after you are in the screen should retrieve your location again for use
Previously it will only get your data when the object was first created. Obviously this won't help if the location is detected wrongly and corrected subsequently. You have to go back and reenter the screen for it to update which is stupid UX. Therefore it now gets it everytime you call

Also settings now have a "Location Value" option in the Dev Settings which will tell you what your coordinates are as detected by the app

Finishes and closes #127