Implemented the auto-refresh rate setting
At this rate we are basically done and can close #124
All that's left are cosmetic changes and maybe adding the arrow into the direction of the bus. But those are cosmetic changes. Lets issue a PR and get Codeclimate running. I think that will throw out heaps of issues tbh xD